feat(stats): move ethereum stats into separate controller and service
