+++ apps/server/src/ethereum/EthereumModule.ts
@@ -9,6 +9,8 @@ import { SemaphoreCache } from '../libs/caches/SemaphoreCache';
 import { EthersModule } from '../modules/EthersModule';
 import { PrismaService } from '../PrismaService';
 import { EthereumController } from './controllers/EthereumController';
+import { StatsController } from './controllers/StatsController';
+import { EthereumStatsService } from './services/EthereumStatsService';
 import { EVMTransactionConfirmerService } from './services/EVMTransactionConfirmerService';
 
 @Module({
@@ -20,10 +22,11 @@ import { EVMTransactionConfirmerService } from './services/EVMTransactionConfirm
     WhaleApiClientProvider,
     DeFiChainTransactionService,
     EVMTransactionConfirmerService,
+    EthereumStatsService,
     SemaphoreCache,
   ],
-  controllers: [EthereumController],
+  controllers: [EthereumController, StatsController],
   imports: [EthersModule, CacheModule.register({ max: 10_000 })],
-  exports: [EVMTransactionConfirmerService],
+  exports: [EVMTransactionConfirmerService, EthereumStatsService],
 })
 export class EthereumModule {}
